+++ integration/spring-data/testsuite/webmvc/src/main/java/com/blazebit/persistence/spring/data/testsuite/webmvc/entity/IdClassEntity.java
@@ -1,0 +1,72 @@
+/*
+ * Copyright 2014 - 2022 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.spring.data.testsuite.webmvc.entity;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.IdClass;
+import java.io.Serializable;
+import java.util.UUID;
+
+/**
+ * @author Christian Beikov
+ * @since 1.6.8
+ */
+@Entity
+@IdClass(IdClassEntityId.class)
+public class IdClassEntity implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    private String id;
+    private String name;
+    private long age;
+
+    public IdClassEntity() {
+    }
+
+    public IdClassEntity(String name) {
+        this.id = UUID.randomUUID().toString();
+        this.name = name;
+    }
+
+    @Id
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    @Id
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public long getAge() {
+        return age;
+    }
+
+    public void setAge(long age) {
+        this.age = age;
+    }
+
+}